Seperate Setting from camera, and make it a new app
