+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/azure_settings/__init__.py
@@ -1,0 +1,4 @@
+"""
+Setting init
+"""
+default_app_config = 'azure_settings.apps.AzureSettingsConfig'
