+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/azure_settings/admin.py
@@ -1,0 +1,9 @@
+"""
+Setting admin
+"""
+from django.contrib import admin
+
+from .models import Setting
+
+# Register your models here.
+admin.site.register(Setting)
