+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/azure_settings/tests/test_model_setting.py
@@ -22,7 +22,7 @@ from azure.cognitiveservices.vision.customvision.training.models import (
     Project)
 
 from cameras.models import Setting
-from config import ENDPOINT, TRAINING_KEY
+from configs.customvision_config import TRAINING_KEY, ENDPOINT
 
 PROJECT_PREFIX = "UnitTest"
