+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/azure_settings/tests/test_special_strings.py
@@ -1,0 +1,7 @@
+"""
+Some special string/char that may cause error
+"""
+special_strings = [
+    "%", "_", "()", "[", "]", "[]", "$", "^", "[^]", "[^a-f]",
+    "`", "&", "?", ".", ",", "'", '"', "#", "--", ";",
+    "+", "-", "*", "/", "{", "}", "{}", "\\"]
