+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/azure_settings/tests/test_template.py
@@ -1,14 +1,15 @@
 """
-Testing Template
+Testing Stram API View
 """
 import logging
 
-# from django.urls import reverse
-# from rest_framework import status
+from django.urls import reverse
+from rest_framework import status
 from rest_framework.test import APITransactionTestCase
-# from configs.customvision_config import ENDPOINT, TRAINING_KEY
-#
-# PROJECT_PREFIX = "UnitTest"
+
+from config import ENDPOINT, TRAINING_KEY
+
+PROJECT_PREFIX = "UnitTest"
 
 logger = logging.getLogger(__name__)
 
@@ -17,14 +18,19 @@ class StreamViewTestCase(APITransactionTestCase):
     """
     Testing stream api view
     """
+
     def setUp(self):
         """
         Create setting, camera, location and parts.
         """
+        pass
+
     def test_setup_is_valid(self):
         """
         Make sure setup is valid
         """
+        pass
+
     def test_valid_setting_list_project(self):
         """
         @Type
@@ -33,6 +39,8 @@ class StreamViewTestCase(APITransactionTestCase):
 
         @Expected Results
         """
+        pass
+
     def test_invalid_setting_list_project(self):
         """
         @Type
@@ -41,12 +49,13 @@ class StreamViewTestCase(APITransactionTestCase):
 
         @Expected Results
         """
+        pass
+
     @classmethod
     def tearDownClass(cls):
-        # trainer = CustomVisionTrainingClient(
-        #     api_key=TRAINING_KEY, endpoint=ENDPOINT)
-        # projects = trainer.get_projects()
-        # for project in projects:
-        #    if project.name.find(PROJECT_PREFIX) == 0:
-        # trainer.delete_project(project_id=project.id)
-        pass
+        trainer = CustomVisionTrainingClient(
+            api_key=TRAINING_KEY, endpoint=ENDPOINT)
+        projects = trainer.get_projects()
+        for project in projects:
+            if project.name.find(PROJECT_PREFIX) == 0:
+                trainer.delete_project(project_id=project.id)
