+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/cameras/admin.py
@@ -1,9 +1,11 @@
+"""
+Camera admin
+"""
 from django.contrib import admin
-from .models import Project, Camera, Image, Part, Setting
+from .models import Project, Camera, Image, Part
 
 # Register your models here.
 admin.site.register(Project)
 admin.site.register(Camera)
 admin.site.register(Image)
 admin.site.register(Part)
-admin.site.register(Setting)
