+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/cameras/apps.py
@@ -1,13 +1,10 @@
 """
-App start
+Camera App start
 """
 import logging
 import sys
 
 from django.apps import AppConfig
-
-from configs.customvision_config import ENDPOINT, TRAINING_KEY
-from configs.iot_config import IOT_HUB_CONNECTION_STRING, DEVICE_ID, MODULE_ID
 
 logger = logging.getLogger(__name__)
 
@@ -27,52 +24,11 @@ class CameraConfig(AppConfig):
         """
         # FIXME test may use this as well
         if 'runserver' in sys.argv:
-            from cameras.models import (Part, Camera, Project, Train, Setting)
-            from locations.models import Location
+            from cameras.models import (Part, Camera, Project, Train)  # pylint: disable=C0415
+            from locations.models import Location  # pylint: disable=C0415
+            from azure_settings.models import Setting  # pylint: disable=C0415
+
             logger.info("CameraAppConfig ready while running server")
-
-            existing_settings = Setting.objects.filter(
-                name=DEFAULT_SETTING_NAME,
-                training_key=TRAINING_KEY,
-                endpoint=ENDPOINT)
-            if len(existing_settings) == 1:
-                logger.info("Found existing %s. Revalidating in pre_save...",
-                            DEFAULT_SETTING_NAME)
-                default_setting = existing_settings[0]
-                default_setting.save()
-
-            elif len(Setting.objects.filter(name=DEFAULT_SETTING_NAME)) > 0:
-                logger.info("Found existing %s with different (Endpoint, key)",
-                            DEFAULT_SETTING_NAME)
-                logger.info("User may already changed the key ")
-                # settings_with_dup_name.delete()
-                default_setting = Setting.objects.filter(
-                    name=DEFAULT_SETTING_NAME)[0]
-                default_setting.save()
-
-            elif len(
-                    Setting.objects.filter(endpoint=ENDPOINT,
-                                           training_key=TRAINING_KEY)) > 0:
-                logger.info(
-                    "Found existing (Endpoint, key) with different setting name"
-                )
-                logger.info("Pass...")
-
-                default_setting = Setting.objects.filter(
-                    endpoint=ENDPOINT, training_key=TRAINING_KEY)[0]
-                default_setting.save()
-            else:
-                logger.info("Creating new %s", DEFAULT_SETTING_NAME)
-                default_setting, created = Setting.objects.update_or_create(
-                    name=DEFAULT_SETTING_NAME,
-                    training_key=TRAINING_KEY,
-                    endpoint=ENDPOINT,
-                    iot_hub_connection_string=IOT_HUB_CONNECTION_STRING,
-                    device_id=DEVICE_ID,
-                    module_id=MODULE_ID)
-                if not created:
-                    logger.error("%s not created. Something went wrong",
-                                 DEFAULT_SETTING_NAME)
 
             create_demo = True
             if create_demo:
@@ -101,7 +57,7 @@ class CameraConfig(AppConfig):
                         'train',
                         'tvmonitor',
                 ]:
-                    demo_part, created = Part.objects.update_or_create(
+                    _, created = Part.objects.update_or_create(
                         name=partname,
                         is_demo=True,
                         defaults={'description': "Demo"})
@@ -116,22 +72,27 @@ class CameraConfig(AppConfig):
                     })
 
                 logger.info("Creating Demo Location")
+
                 # Demo Location should be created already
-                # FIXME : split location from project, and remove location here
+                # FIXME split location from project, and remove location here
                 demo_locations = Location.objects.filter(is_demo=True)
                 if len(demo_locations) <= 0:
+                    return
+                default_settings = Setting.objects.filter(
+                    name=DEFAULT_SETTING_NAME)
+                if len(default_settings) <= 0:
                     return
 
                 logger.info("Creating Demo Project")
                 demo_project, created = Project.objects.update_or_create(
                     is_demo=True,
                     defaults={
-                        'setting': default_setting,
+                        'setting': default_settings.first(),
                         'camera': demo_camera,
                         'location': demo_locations.first(),
                     })
 
-                demo_train, created = Train.objects.update_or_create(
+                _, created = Train.objects.update_or_create(
                     project=demo_project,
                     defaults={
                         'status': 'demo ok',
@@ -143,8 +104,9 @@ class CameraConfig(AppConfig):
             if len(demo_locations) <= 0:
                 return
 
-            # FIXME: is DEMO = False, will not have camera, location to assign
-            default_project, created = Project.objects.update_or_create(
+            # Create default project
+            # FIXME is DEMO = False, will not have camera, location to assign
+            _, created = Project.objects.update_or_create(
                 is_demo=False,
                 defaults={
                     'camera': demo_camera,
