+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/cameras/serializers.py
@@ -7,8 +7,7 @@ import logging
 from django.db.utils import IntegrityError
 from rest_framework import serializers
 
-from .models import (Annotation, Camera, Image, Part, Project, Setting, Task,
-                     Train)
+from .models import (Annotation, Camera, Image, Part, Project, Task, Train)
 
 logger = logging.getLogger(__name__)
 
@@ -74,43 +73,6 @@ class TaskSerializer(serializers.HyperlinkedModelSerializer):
     class Meta:
         model = Task
         fields = ["task_type", "status", "log", "project"]
-
-
-class SettingSerializer(serializers.HyperlinkedModelSerializer):
-    """SettingSerializer"""
-    class Meta:
-        model = Setting
-        fields = [
-            "id",
-            "name",
-            "training_key",
-            "endpoint",
-            "is_trainer_valid",
-            "iot_hub_connection_string",
-            "device_id",
-            "module_id",
-            "is_collect_data",
-            "obj_detection_domain_id",
-        ]
-
-    def create(self, validated_data):
-        obj, _ = Setting.objects.get_or_create(
-            endpoint=validated_data["endpoint"],
-            training_key=validated_data["training_key"],
-            defaults={
-                "name":
-                validated_data["name"],
-                "iot_hub_connection_string":
-                validated_data["iot_hub_connection_string"],
-                "device_id":
-                validated_data["device_id"],
-                "module_id":
-                validated_data["module_id"],
-                "is_collect_data":
-                validated_data["is_collect_data"],
-            },
-        )
-        return obj
 
 
 class ProjectSerializer(serializers.HyperlinkedModelSerializer):
