+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/cameras/views.py
@@ -47,18 +47,17 @@ from .models import (
     Image,
     Part,
     Project,
-    Setting,
     Stream,
     Task,
     Train,
 )
+
 from .serializers import (
     AnnotationSerializer,
     CameraSerializer,
     ImageSerializer,
     PartSerializer,
     ProjectSerializer,
-    SettingSerializer,
     TaskSerializer,
     TrainSerializer,
 )
@@ -356,82 +355,6 @@ class TaskViewSet(FiltersMixin, viewsets.ModelViewSet):
         "project": "project",
     }
 
-
-class SettingViewSet(viewsets.ModelViewSet):
-    """
-    Setting ModelViewSet
-    """
-
-    queryset = Setting.objects.all()
-    serializer_class = SettingSerializer
-
-    @action(detail=True, methods=["get"])
-    def list_projects(self, request, pk=None):
-        """
-        List Project under Training Key + Endpoint
-        """
-        try:
-            setting_obj = Setting.objects.get(pk=pk)
-            if not setting_obj.training_key:
-                raise ValueError("Training Key")
-            if not setting_obj.endpoint:
-                raise ValueError("Endpoint")
-            trainer = setting_obj.get_trainer_obj()
-            result = {}
-            project_list = trainer.get_projects()
-            for project in project_list:
-                result[project.id] = project.name
-            return JsonResponse(result)
-        except ValueError:
-            return JsonResponse(
-                {
-                    "status": "failed",
-                    "log": error_messages.CUSTOM_VISION_MISSING_FIELD
-                },
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-        except KeyError as key_err:
-            if str(key_err) in ["Endpoint", "'Endpoint'"]:
-                return JsonResponse(
-                    {
-                        "status": "failed",
-                        "log": error_messages.CUSTOM_VISION_ACCESS_ERROR,
-                    },
-                    status=status.HTTP_400_BAD_REQUEST,
-                )
-            return JsonResponse(
-                {
-                    "status": "failed",
-                    "log": f"KeyError {str(key_err)}"
-                },
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            )
-        except CustomVisionErrorException as customvision_error:
-            if (customvision_error.message ==
-                    "Operation returned an invalid status code 'Access Denied'"
-                ):
-
-                return JsonResponse(
-                    {
-                        "status": "failed",
-                        "log": error_messages.CUSTOM_VISION_ACCESS_ERROR,
-                    },
-                    status=status.HTTP_503_SERVICE_UNAVAILABLE,
-                )
-            return JsonResponse(
-                {
-                    "status": "failed",
-                    "log": customvision_error.message
-                },
-                status=customvision_error.response.status_code,
-            )
-        except Exception as e:
-            logger.exception("Unexpected Error while listing projects")
-            return JsonResponse({
-                "status": "failed",
-                "log": str(e)
-            },
-                                status=status.HTTP_400_BAD_REQUEST)
 
 
 class ProjectViewSet(FiltersMixin, viewsets.ModelViewSet):
